move logic to general archive handler.
