+++ pkg/handlers/handlers.go
@@ -1,16 +1,9 @@
 package handlers
 
 import (
-	"bytes"
 	"context"
-	"fmt"
 	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 
-	logContext "github.com/trufflesecurity/trufflehog/v3/pkg/context"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/sources"
 )
 
@@ -20,39 +13,59 @@ func DefaultHandlers() []Handler {
 	}
 }
 
+// SpecializedHandler defines the interface for handlers that can process specialized archives.
+// It includes a method to handle specialized archives and determine if the file is of a special type.
+type SpecializedHandler interface {
+	// HandleSpecialized examines the provided file reader within the context and determines if it is a specialized archive.
+	// It returns a reader with any necessary modifications, a boolean indicating if the file was specialized,
+	// and an error if something went wrong during processing.
+	HandleSpecialized(context.Context, io.Reader) (io.Reader, bool, error)
+}
+
 type Handler interface {
 	FromFile(context.Context, io.Reader) chan ([]byte)
 	IsFiletype(context.Context, io.Reader) (io.Reader, bool)
 	New()
 }
 
+// HandleFile processes a given file by selecting an appropriate handler from DefaultHandlers.
+// It first checks if the handler implements SpecializedHandler for any special processing,
+// then falls back to regular file type handling. If successful, it reads the file in chunks,
+// packages them in the provided chunk skeleton, and sends them to chunksChan.
+// The function returns true if processing was successful and false otherwise.
+// Context is used for cancellation, and the caller is responsible for canceling it if needed.
 func HandleFile(ctx context.Context, file io.Reader, chunkSkel *sources.Chunk, chunksChan chan *sources.Chunk) bool {
-	// Find a handler for this file.
-	var handler Handler
 	for _, h := range DefaultHandlers() {
 		h.New()
+		var (
+			isSpecial bool
+			err       error
+		)
+
+		// Check if the handler implements SpecializedHandler and process accordingly.
+		if specialHandler, ok := h.(SpecializedHandler); ok {
+			if file, isSpecial, err = specialHandler.HandleSpecialized(ctx, file); isSpecial && err == nil {
+				return handleChunks(ctx, h.FromFile(ctx, file), chunkSkel, chunksChan)
+			}
+		}
+
 		var isType bool
 		if file, isType = h.IsFiletype(ctx, file); isType {
-			handler = h
-			break
+			return handleChunks(ctx, h.FromFile(ctx, file), chunkSkel, chunksChan)
 		}
 	}
-	if handler == nil {
-		return false
-	}
+	return false
+}
 
-	// Process the file and read all []byte chunks from handlerChan.
-	handlerChan := handler.FromFile(ctx, file)
+func handleChunks(ctx context.Context, handlerChan chan []byte, chunkSkel *sources.Chunk, chunksChan chan *sources.Chunk) bool {
 	for {
 		select {
 		case data, open := <-handlerChan:
 			if !open {
-				// We finished reading everything from handlerChan.
 				return true
 			}
 			chunk := *chunkSkel
 			chunk.Data = data
-			// Send data on chunksChan.
 			select {
 			case chunksChan <- &chunk:
 			case <-ctx.Done():
@@ -62,152 +75,4 @@ func HandleFile(ctx context.Context, file io.Reader, chunkSkel *sources.Chunk, c
 			return false
 		}
 	}
-}
-
-const (
-	debFileExtension = ".deb"
-	rpmFileExtension = ".rpm"
-)
-
-// HandleSpecializedArchives takes a file path and an io.ReadCloser representing the input file,
-// and processes it based on its extension, such as handling Debian (.deb) and RPM (.rpm) packages.
-// It returns an io.ReadCloser that can be used to read the processed content of the file,
-// and an error if any issues occurred during processing.
-// The caller is responsible for closing the returned reader.
-func HandleSpecializedArchives(ctx logContext.Context, path string, inputFile io.ReadCloser) (io.ReadCloser, error) {
-	var reader io.ReadCloser
-	var err error
-	ext := filepath.Ext(path)
-	switch ext {
-	case debFileExtension:
-		reader, err = extractDebContent(ctx, inputFile)
-	case rpmFileExtension:
-		reader, err = extractRpmContent(ctx, inputFile)
-	default:
-		reader = inputFile
-	}
-	if err != nil {
-		return nil, fmt.Errorf("unable to extract file with extension %s: %w", ext, err)
-	}
-	return reader, nil
-}
-
-// extractDebContent takes a .deb file as an io.ReadCloser, extracts its contents
-// into a temporary directory, and returns a ReadCloser for the extracted data archive.
-// It handles the extraction process by using the 'ar' command and manages temporary
-// files and directories for the operation.
-// The caller is responsible for closing the returned reader.
-func extractDebContent(_ logContext.Context, file io.ReadCloser) (io.ReadCloser, error) {
-	tempEnv, err := createTempEnv(file)
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(tempEnv.tempFileName)
-	defer os.RemoveAll(tempEnv.extractPath)
-
-	cmd := exec.Command("ar", "x", tempEnv.tempFile.Name())
-	cmd.Dir = tempEnv.extractPath
-	if err := executeCommand(cmd); err != nil {
-		return nil, err
-	}
-
-	// List the content of the extraction directory.
-	extractedFiles, err := os.ReadDir(tempEnv.extractPath)
-	if err != nil {
-		return nil, fmt.Errorf("unable to read extracted directory: %w", err)
-	}
-
-	// Determine the correct data archive name. (e.g., data.tar.gz, data.tar.xz)
-	var dataArchiveName string
-	for _, file := range extractedFiles {
-		if strings.HasPrefix(file.Name(), "data.tar.") {
-			dataArchiveName = file.Name() // Use the actual name if different
-			break
-		}
-	}
-
-	return openDataArchive(tempEnv.extractPath, dataArchiveName)
-}
-
-// extractRpmContent takes an .rpm file as an io.ReadCloser, extracts its contents
-// into a temporary directory, and returns a ReadCloser for the extracted data archive.
-// It handles the extraction process by using the 'rpm2cpio' and 'cpio' commands and manages temporary
-// files and directories for the operation.
-// The caller is responsible for closing the returned reader.
-func extractRpmContent(ctx logContext.Context, file io.ReadCloser) (io.ReadCloser, error) {
-	tempEnv, err := createTempEnv(file)
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(tempEnv.tempFileName)
-	defer os.RemoveAll(tempEnv.extractPath)
-
-	// Use rpm2cpio to convert the RPM file to a cpio archive and then extract it using cpio command.
-	cmd := exec.Command("sh", "-c", "rpm2cpio "+tempEnv.tempFile.Name()+" | cpio -id")
-	cmd.Dir = tempEnv.extractPath
-	if err := executeCommand(cmd); err != nil {
-		return nil, err
-	}
-
-	// List the content of the extraction directory.
-	extractedFiles, err := os.ReadDir(tempEnv.extractPath)
-	if err != nil {
-		return nil, fmt.Errorf("unable to read extracted directory: %w", err)
-	}
-
-	var dataArchiveName string
-	// Determine the correct data archive name.
-	for _, file := range extractedFiles {
-		if strings.HasSuffix(file.Name(), ".tar.gz") {
-			dataArchiveName = file.Name()
-			break
-		}
-	}
-
-	return openDataArchive(tempEnv.extractPath, dataArchiveName)
-}
-
-type tempEnv struct {
-	tempFile     *os.File
-	tempFileName string
-	extractPath  string
-}
-
-// createTempEnv creates a temporary file and a temporary directory for extracting archives.
-// The caller is responsible for removing these temporary resources
-// (both the file and directory) when they are no longer needed.
-func createTempEnv(file io.ReadCloser) (tempEnv, error) {
-	tempFile, err := os.CreateTemp("", "tmp")
-	if err != nil {
-		return tempEnv{}, fmt.Errorf("unable to create temporary file: %w", err)
-	}
-
-	extractPath, err := os.MkdirTemp("", "tmp_archive")
-	if err != nil {
-		return tempEnv{}, fmt.Errorf("unable to create temporary directory: %w", err)
-	}
-
-	if _, err = io.Copy(tempFile, file); err != nil {
-		return tempEnv{}, fmt.Errorf("unable to copy content to temporary file: %w", err)
-	}
-
-	return tempEnv{tempFile: tempFile, tempFileName: tempFile.Name(), extractPath: extractPath}, nil
-}
-
-func executeCommand(cmd *exec.Cmd) error {
-	var stderr bytes.Buffer
-	cmd.Stderr = &stderr
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("unable to execute command: %w; error: %s", err, stderr.String())
-	}
-	return nil
-}
-
-func openDataArchive(extractPath string, dataArchiveName string) (io.ReadCloser, error) {
-	dataArchivePath := filepath.Join(extractPath, dataArchiveName)
-	dataFile, err := os.Open(dataArchivePath)
-	if err != nil {
-		return nil, fmt.Errorf("unable to open file: %w", err)
-	}
-	return dataFile, nil
 }
