+++ pkg/sources/filesystem/filesystem.go
@@ -151,13 +151,7 @@ func (s *Source) scanFile(ctx context.Context, path string, chunksChan chan *sou
 	defer inputFile.Close()
 	logger.V(3).Info("scanning file")
 
-	reader, err := handlers.HandleSpecializedArchives(ctx, path, inputFile)
-	if err != nil {
-		return fmt.Errorf("unable to handle specialized archives: %w", err)
-	}
-	defer reader.Close()
-
-	reReader, err := diskbufferreader.New(reader)
+	reReader, err := diskbufferreader.New(inputFile)
 	if err != nil {
 		return fmt.Errorf("could not create re-readable reader: %w", err)
 	}
